+++ keyboards/rk/b87/config.h
@@ -41,39 +41,16 @@
 #define RGB_MATRIX_KEYPRESSES
 
 /* RGB Light */
-#define RGBLED_NUM         8
-#define RGBLIGHT_LED_COUNT RGBLED_NUM
+#define RGBLED_NUM 8
 #define RGBLIGHT_CUSTOM_DRIVER
-// #define RGBLIGHT_LED_MAP { 0,1,2,3,4,5,6,7 }
 
-/* RGB Lighting effects*/
-#define RGBLIGHT_ANIMATIONS
-#define RGBLIGHT_EFFECT_ALTERNATING
-#define RGBLIGHT_EFFECT_BREATHING
-#define RGBLIGHT_EFFECT_CHRISTMAS
-#define RGBLIGHT_EFFECT_KNIGHT
-#define RGBLIGHT_EFFECT_RAINBOW_MOOD
-#define RGBLIGHT_EFFECT_RAINBOW_SWIRL
-#define RGBLIGHT_EFFECT_RGB_TEST
-#define RGBLIGHT_EFFECT_SNAKE
-#define RGBLIGHT_EFFECT_STATIC_GRADIENT
-#define RGBLIGHT_EFFECT_TWINKLE
-#define RGBLIGHT_DISABLE_KEYCODES
-
-#define CUSTOM_RGBLIGHT_EFFECT_BREATHE_TABLE
-#define CUSTOM_RGBLIGHT_EFFECT_RAINBOW_MOOD
-#define CUSTOM_RGBLIGHT_EFFECT_RAINBOW_SWIRL
-
-#define RGBLIGHT_DEFAULT_VAL 60
-#define breathing_val        40
-#define rainbow_mood         40
-#define RGBLIGHT_SLEEP
-#define RGBLIGHT_DEFAULT_MODE   14
+#define breathing_val           40
+#define rainbow_mood            40
 #define LED_TYPE                RGB
+
 /* WS2812 */
 #define WS2812_SPI_DRIVER       SPIDM2
 #define WS2812_SPI_DIVISOR      32
-#define RGB_MATRIX_DEFAULT_MODE 13
 
 /* rgb_record */
 #define ENABLE_RGB_MATRIX_RGBR_PLAY
