+++ keyboards/rk/b87/rgb_record/rgb_record.mk
@@ -1,2 +1,3 @@
 SRC += rgb_record/rgb_record.c
 SRC += rgb_record/rgb_rgblight.c
+SRC += rgb_record/rgb_effect.c
